Need app+REST to server config JSON for videos
Stand up an Erlang application and a Cowboy web server to server up a
diagram/map of video URLs to "How To" Erlang videos.

The goal is to have a web page with video URL squares that I can move
around. Videos near to each other are more related; videos near the top
are more basic. I'd like to be able to add new videos, edit the URL,
and edit the description and CSS attributes of the URLs (e.g. border).
Whenever a URL is moved or edited, the web page JavaScript will post a
new URL video data JSON payload to the web server, which will write the
new JSON to file.

The web page will download the video URL data from a REST endpoint
running on the web server.