Web page needs static assets and JSON
Moved the static assets to the priv/static folder and created static
routes to the virtual "assets/" folder.

Put the URLs JSON in a file, served from the priv directory.

Moved the URLs REST handler to video_urls.erl

Added Jiffy as a dependency, although I'm not sure I'm going to need it,
since I'm just storing the JSON in a file and serving it.

Added the ability to read JSON in the JavaScript with the new "fetch()"
gear, that uses promises. Apparently, you can "await" a promise within
an "async" function, or you can use .then(goodFun, badFun) promise
resolution function. I just used the example code from:
https://developer.mozilla.org/en-US/docs/Web/API/Fetch_API/Using_Fetch