Draw URLs from urls.json as divs
"fixed" means the positions of the divs are relative the browser window.
"absolute" seems to mean that the div would be positioned relative to
the first non-"static" (default) parent element, called the first
"positioned" parent element.

The browser or the server seems to be caching the .js file, so I had to
restart it to get the updated JavaScript. I didn't see anywhere in the
Cowboy docs to change this.