Clicking and dragging should move URLs
If you drag an element fast enough to get the mouse cursor outside of
the element and then mouseup, the element won't get the mouseup event.
Use the document to capture the mousemouve and mouseup events.

You might still be able to move the mouse cursor outside of the
document and mouseup. Oh well.