Dbl-click to add URLs, working on resizing
Working on adding resize handles to each URL box; I'm rejiggering the
drag and drop handlers so that they can be reused. Each subscriber to
drag and drop needs to have two lists of handlers: move handlers and up
handlers; move is for when the mouse moves and up is for when the mouse
button is released.

This is still a work in progress.

Many thanks to the streamer who helped me out today but didn't want any
credit! I learned a lot of cool JavaScript stuff.

TODO:
- when a URL is moved, moved the resize handles
- when a resize handle is moved resize the URL box and also move it
  if the resize puts the top or left sides go down or right.
  (Or else resizing the topleft down and right would cause the bottom
   right to move close to the top left.)
- change mouse cursor when hovering over resize handles
- hide resize handles