Nice rounded corners, less aggressive handles
Rounding the corners required some padding for the content and adjusting
where the resize handles sit.

The CSS top and left are at the top left of the border, outside the
padding, whereas the CSS width and height, are the width and height of
the content inside the padding. To go from left border to right border,
we need left + padding + width + padding.

TODO:
- fix the resize handle coordinates when the URL box moves.