Allow resize of non-clamped dimension
If a dimension of a URL box (e.g. height, width) is at the minimum,
allow the other dimension to still be modified. Previously, if either
dimension was at the minimum the resizing froze until the offending
dimension was increased. Now, only a dimension that is at the minimum is
prevented from getting smaller, the other dimension can be modified.

Also did some refactoring.

Note that there's a difference between "offsetWidth" and "style.width":
offsetWidth includes the "content" box and the padding on both sides ...
and maybe the border, whereas style.width only includes the content box,
without padding or border.