Resize boxes need to be centered on corners
I was able to use the "offset to mouse event" logic used in the main
div, when it's being dragged, to position the resize divs.

Spam the REPL less while Cowboy is handling requests.

Simplify saving the URL box coordinates to the urlData when the URL box
is moved. I keep forgetting that element.style.left is a string, e.g.
'150px'; it requires conversion to an int, which conveniently drops
anything after the first non-digit character.